refactor: /cart에 대한 View를 설정 파일에서의 viewController로 처리되도록 변경
