--- src/main/java/cart/web/cart/controller/CartController.java
@@ -1,13 +1,0 @@
-package cart.web.cart.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class CartController {
-
-    @GetMapping("/cart")
-    public String showCart() {
-        return "cart";
-    }
-}
