+++ src/test/java/cart/web/ViewControllerRegistryTest.java
@@ -1,4 +1,4 @@
-package cart.web.cart.controller;
+package cart.web;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayNameGeneration;
@@ -16,7 +16,8 @@ import io.restassured.RestAssured;
 @Sql(scripts = "classpath:schema-truncate.sql", executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD)
 @SuppressWarnings("NonAsciiCharacters")
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
-class CartControllerTest {
+class ViewControllerRegistryTest {
+
     @LocalServerPort
     private int port;
